BAS-73: generalise the PA-level model from K% to any per-PA rate
src/models/pa_rate.py is src/models/pa_k_rate.py with the three things that
actually differ between K%, BB% and HR/PA lifted into a RateComponent: the
numerator column (is_k / is_bb / is_hr), the league_init prior mean, and the
direction of the age curve. Everything else — the cells, the Binomial, the
league walk, the ability walk, the pitcher effect, the park and handedness
terms — is one copy serving all three.

bb_rate counts is_bb alone, not is_bb + is_hbp, matching what
src.eval.backtest.COMPONENTS["bb_rate"] scores (intraseason.aggregate_pa
builds `bb` from is_bb and keeps `hbp` separate). HBP is ~1% of PA against
BB's ~8.5%: big enough to matter, small enough to miss.

league_init's prior mean is now read off the training data's earliest season
— the season the league random walk actually starts at — instead of a
hard-coded number: about -2.38 for BB% (~8.5%) and -3.44 for HR/PA (~3.1%)
on 2019-2026 data. k_rate is the one component that keeps its pin at -1.27,
because that is what the published sweep was fit under and pinning it is what
makes the bit-for-bit check below possible. The cost of the pin is small: the
data-driven value for K% runs -1.21 to -1.26 across those seasons, a fifth of
the prior's own sd of 0.3.

The age curve follows src.eval.tuning.AGE_DIRECTION per component. K% is the
one component where a bigger rate is a worse hitter (-1), so its constrained
curve is a valley — falling toward the peak, rising after, the shape the
sweep's fitted slope_old = 0.0080 describes. BB% and HR/PA (+1) are hills.
The free quadratic's beta_age2 prior mean flips the same way (+0.005 / -0.005)
while leaving the sign to the data. The projection side reapplies the same
sign in _age_term_function, since the trace stores the slopes as the positive
HalfNormals they are.

pa_k_rate.py is now a thin wrapper that pins component="k_rate", and K% being
unchanged is a checked claim rather than an assertion:
tests/test_models/test_pa_rate.py fits a fixture through it in all four
ModelOptions combinations and compares every posterior mean and every
projected rate against tests/test_models/k_rate_reference.json — generated
from the pre-split module, compared with no tolerance at all. The fixture
samples with the pure-Python pymc NUTS backend at a fixed seed precisely
because that reproduces the same draws in a fresh process; verified across
two processes before being frozen. All four combos match exactly.
