+++ tests/test_models/k_rate_reference.json
@@ -1,0 +1,182 @@
+{
+ "age": {
+  "batters": [
+   1000,
+   1009,
+   1008,
+   1011,
+   1001,
+   1003,
+   1002,
+   1007,
+   1004,
+   1010,
+   1006,
+   1005
+  ],
+  "posterior_means": {
+   "beta_hand": 0.06135367175177667,
+   "league_init": -1.225938714412826,
+   "league_innovations": -0.004475346308523518,
+   "league_trend": -1.2393551658613058,
+   "mu_ability": 0.1198653418055149,
+   "park_effect": 0.0,
+   "peak_age": 28.74220815714196,
+   "peak_frac": 0.623701359523659,
+   "player_ability": 0.0950663722908033,
+   "sigma_ability": 0.2796024251369873,
+   "slope_old": 0.015825590903355332,
+   "slope_young": 0.020848023458388258,
+   "z_ability": -0.0994260610312266
+  },
+  "projected_k_rate": [
+   0.18376602661068195,
+   0.217901777477693,
+   0.2249559425358828,
+   0.22694882855350493,
+   0.22721419449848768,
+   0.2376612623765095,
+   0.26547251416055134,
+   0.27981538296238323,
+   0.2873584760466043,
+   0.29712299410658016,
+   0.324875523147863,
+   0.3473249052174674
+  ]
+ },
+ "flat": {
+  "batters": [
+   1000,
+   1009,
+   1001,
+   1008,
+   1011,
+   1003,
+   1002,
+   1007,
+   1004,
+   1010,
+   1006,
+   1005
+  ],
+  "posterior_means": {
+   "beta_age": -0.009566283804743286,
+   "beta_age2": 0.009966680273021122,
+   "beta_hand": 0.016162832345778692,
+   "league_init": -1.2516251639790026,
+   "league_innovations": 0.00733929507870422,
+   "league_trend": -1.232011806336834,
+   "mu_ability": 0.040222922094982766,
+   "park_effect": 0.0,
+   "player_ability": 0.08095206854722217,
+   "sigma_ability": 0.2798125582273666,
+   "z_ability": 0.13193201012602646
+  },
+  "projected_k_rate": [
+   0.18005623121156972,
+   0.21121878047354425,
+   0.21529428772921683,
+   0.22007271426180577,
+   0.22046947407994033,
+   0.2238616368461408,
+   0.25609050548942197,
+   0.2629421143878224,
+   0.2817790317775104,
+   0.2876430996826233,
+   0.3187900027158622,
+   0.32657118420668807
+  ]
+ },
+ "walk": {
+  "batters": [
+   1000,
+   1009,
+   1011,
+   1001,
+   1008,
+   1003,
+   1002,
+   1007,
+   1004,
+   1010,
+   1006,
+   1005
+  ],
+  "posterior_means": {
+   "beta_age": -0.005828164617788108,
+   "beta_age2": 0.008713028190313471,
+   "beta_hand": 0.0020652119192991073,
+   "league_init": -1.1713939470002055,
+   "league_innovations": 0.00024619275126957027,
+   "league_trend": -1.1653948020035196,
+   "mu_ability": 0.05058645235881934,
+   "park_effect": 0.0,
+   "player_ability": 0.047011196377197736,
+   "sigma_ability": 0.300980902695365,
+   "sigma_step": 0.08961243712647061,
+   "z_ability": -0.055790522738440484,
+   "z_step": -0.0006750472466523376
+  },
+  "projected_k_rate": [
+   0.17539051260503108,
+   0.20606449171253924,
+   0.2084002065017664,
+   0.216806251454983,
+   0.21947199957183655,
+   0.2470904066170012,
+   0.2577064549277913,
+   0.26713896377869717,
+   0.2799300743671384,
+   0.3019860744763191,
+   0.32513935837568697,
+   0.3384712455386204
+  ]
+ },
+ "walk_age": {
+  "batters": [
+   1000,
+   1009,
+   1011,
+   1008,
+   1001,
+   1003,
+   1002,
+   1007,
+   1004,
+   1010,
+   1006,
+   1005
+  ],
+  "posterior_means": {
+   "beta_hand": 0.042359830387680326,
+   "league_init": -1.1839989434376834,
+   "league_innovations": -0.012802974259877364,
+   "league_trend": -1.2020099441665077,
+   "mu_ability": 0.06741079917729324,
+   "park_effect": 0.0,
+   "peak_age": 28.196377309028076,
+   "peak_frac": 0.5327295515046793,
+   "player_ability": 0.07635130097692655,
+   "sigma_ability": 0.31280803649922706,
+   "sigma_step": 0.05916129954678973,
+   "slope_old": 0.013801100734741787,
+   "slope_young": 0.021043564864684554,
+   "z_ability": 0.03820882780798106,
+   "z_step": -0.06666823847540283
+  },
+  "projected_k_rate": [
+   0.17606212828808562,
+   0.21296457492228213,
+   0.21569065481610525,
+   0.2182145178256841,
+   0.2242083451656049,
+   0.24475069824289344,
+   0.2619786020015342,
+   0.26679742788995414,
+   0.28624509182291036,
+   0.2961298273183515,
+   0.3220077597656859,
+   0.34012073086979905
+  ]
+ }
+}
